Initial Gym Wrapper support, Restructured for relative imports. Added a few lines in the readme
