+++ README.md
@@ -1,2 +1,13 @@
 # AandA
 A simple A&amp;A engine. 
+
+# Install
+```bash
+pip install git+https://github.com/cair/Axis_and_Allies.git
+```
+
+# Running Examples
+```bash
+python -m axis_and_allies.demostration_game
+python -m axis_and_allies.example
+```
