+++ axis_and_allies/buildings.py
@@ -1,0 +1,9 @@
+class Buildings:
+
+    class Industry(object):
+        def __init__(self, owner):
+            self.owner = owner
+            self.cost = 10
+
+        def __repr__(self):
+            return "Ind"
