+++ axis_and_allies/map_generator/map_class.py
@@ -1,24 +1,5 @@
 import numpy as np
-
-
-class Tile(object):
-    def __init__(self, cords, name='nameless', water=False):
-        self.owner = None
-
-        self.neighbours = []
-
-        self.cords = cords
-
-        self.units = []#dict()
-
-        self.constructions = []
-
-        self.water = water
-
-        self.value = 2
-
-    def __repr__(self):
-        return self.owner.name+self.units.__str__()+self.constructions.__str__()
+from .tile import Tile
 
 
 class MapClass(object):
