+++ axis_and_allies/new_bot.py
