--- buildings.py
@@ -1,7 +1,0 @@
-class Industry(object):
-    def __init__(self, owner):
-        self.owner = owner
-        self.cost = 10
-
-    def __repr__(self):
-        return "Ind"
